mqtt_docker/python_server/requirements.txt
Flask
Flask-MQTT
aiofiles==24.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.1
crypto==1.4.1
cryptography==43.0.3
cycler==0.12.1
fonttools==4.55.3
idna==3.10
kiwisolver==1.4.7
matplotlib==3.10.0
Naked==0.1.32
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
pycryptodome==3.21.0
pyparsing==3.2.0
pyTelegramBotAPI==4.25.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
shellescape==3.8.1
simonspeckciphers==1.0.0
six==1.17.0
tzdata==2024.2
urllib3==2.2.3
